refactor(ui): extract project workspace shell
